% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ft_download_model}
\alias{ft_download_model}
\title{Download Fasttext Model Files}
\usage{
ft_download_model(lang = "en", mb = 500)
}
\arguments{
\item{lang}{the two letter language code
specifying the language you would
like to download. See the function
\code{\link{ft_languages}} for a
complete list of available choices.}

\item{mb}{the number of megabytes to download
from the file. The default (500) gets
around 200k rows. Adjust as needed.
Set to \code{Inf} to get all rows.
This is a 6GB file.}
}
\value{
Invisibly returns the status code of the download.
       The embedding matrix is stored on disk.
}
\description{
To run embeddings with a specific language, you
must first download the respective language file.
This needs to be done only once. As the raw files
are quite large (over 6GB), this function grabs
only the most frequent words. You can control the
number of words by setting the number of megabytes
to download.
}
\examples{
\dontrun{
ft_download_model(lang = "zh", mb = 200)
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
