% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ft_embed}
\alias{ft_embed}
\title{Word Embeddings}
\usage{
ft_embed(words, lang = "en")
}
\arguments{
\item{words}{a character vector of the words
for which an embedding should be
returned}

\item{lang}{the two letter language code
specifying the language to use for
the embedding}
}
\value{
A matrix with three hundred columns and one row
        for each word. Rownames record the input words.
        Rows corresponding the words not in the embedding
        are set to \code{NA}
}
\description{
Given a vector words, this function returns the
300-dimensional embedding of each element. The
embedding is given in reference to a particular
language. The model must have been loaded
using the function \code{\link{ft_load}}.
}
\examples{
ft_load(lang = "en")
ft_embed(c("the", "and", "I", "banana"), lang = "en")

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
