% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ft_languages}
\alias{ft_languages}
\title{Available Languages}
\usage{
ft_languages()
}
\value{
A data frame of available languages, along with
        information about which ones are already downloaded.
}
\description{
This function prints a data frame of the available languages,
language codes, and which ones are alread downloaded.
}
\examples{

ft_languages()

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
