% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ft_load_model}
\alias{ft_load_model}
\title{Load Fasttext Model Files}
\usage{
ft_load_model(lang = "en")
}
\arguments{
\item{lang}{the two letter language code
specifying the language you would
like to download. See the function
\code{\link{ft_languages}} for a
complete list of available choices.}
}
\description{
Loads the model files into memory for the specified
language. You must first download the respective
language file using \code{\link{ft_download_model}}.
}
\examples{
ft_load_model(lang = "en")

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
