% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ft_nn}
\alias{ft_nn}
\title{Nearest Neighbors}
\usage{
ft_nn(words, lang = "en", lang_out = lang, nn = 10L)
}
\arguments{
\item{words}{a character vector of the words
for which an embedding should be
returned}

\item{lang}{the two letter language code
specifying the language to use for
the embedding of the inputs}

\item{lang_out}{the two letter language code
specifying the language to use for
the embedding of the outputs.}

\item{nn}{positive integer. The number of
nearest neighbors to return in the output.}
}
\value{
A matrix with \code{nn} columns and one row for each
        input word. Rows associated with input words not found
        in the embedding are set to \code{NA}.
}
\description{
Given a vector words, this function returns the
nearest neighbors in the vector space. Users may
select a different language for the nearest neighbors
than is used in the input embedding.
}
\examples{
ft_load(lang = "en")
#ft_load(lang = "fr")
#ft_nn(c("the", "and", "I"), lang = "en", lang_out = "fr")

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
